AB-1413: State locking with identity tx

Closes AB-1413

See merge request alphabill/alphabill!869

+++ txsystem/money/money_tx_system_test.go
@@ -17,6 +17,7 @@ import (
 	"github.com/alphabill-org/alphabill/txsystem/fc"
 	"github.com/alphabill-org/alphabill/txsystem/fc/testutils"
 	"github.com/alphabill-org/alphabill/txsystem/fc/transactions"
+
 	"github.com/alphabill-org/alphabill/txsystem/fc/unit"
 	testtransaction "github.com/alphabill-org/alphabill/txsystem/testutils/transaction"
 	"github.com/alphabill-org/alphabill/types"
@@ -704,7 +705,7 @@ func TestExecute_FeeCreditSequence_OK(t *testing.T) {
 	require.NotNil(t, sm)
 
 	// verify user fee credit is 18 (transfer 20 minus fee 2 * fee)
-	remainingValue := txAmount - (2 * txFee) // 18
+	remainingValue := txAmount - (2 * uint64(txFee)) // 18
 	fcrUnit, err := rmaTree.GetUnit(fcrUnitID, false)
 	require.NoError(t, err)
 	fcrUnitData, ok := fcrUnit.Data().(*unit.FeeCreditRecord)
@@ -773,7 +774,7 @@ func TestExecute_FeeCreditSequence_OK(t *testing.T) {
 	ib, err = rmaTree.GetUnit(initialBill.ID, false)
 	require.NoError(t, err)
 	require.True(t, ok)
-	require.EqualValues(t, initialBill.Value-4*txFee, ib.Data().SummaryValueInput())
+	require.EqualValues(t, initialBill.Value-4*uint64(txFee), ib.Data().SummaryValueInput())
 
 	// and initial bill got unlocked
 	bd, ok = ib.Data().(*BillData)
